Update SignInTest.java
Since we used Singleton Design pattern therefore only one instance created due to which overhead reduces. As you can see i have given method as getInstanceOfDriverManager ,if we invoke this method it will check the object instance is created or not and if it is not created then it will create instances of webdriver or if the instance is already created then we will use that already created instance.